Move banchmarks to gym subvfolder
