+++ library/pyrefly-baseline.json
@@ -5,31 +5,29 @@
       "column": 35,
       "stop_line": 158,
       "stop_column": 54,
-      "path": "src/physicalai/benchmark/benchmark.py",
+      "path": "src/physicalai/benchmark/gyms/benchmark.py",
       "code": -2,
       "name": "missing-argument",
-      "description": "Missing argument `overall_success_rate` in function `physicalai.benchmark.results.BenchmarkResults.__init__`",
-      "concise_description": "Missing argument `overall_success_rate` in function `physicalai.benchmark.results.BenchmarkResults.__init__`",
-      "severity": "error"
+      "description": "Missing argument `overall_success_rate` in function `physicalai.benchmark.gyms.results.BenchmarkResults.__init__`",
+      "concise_description": "Missing argument `overall_success_rate` in function `physicalai.benchmark.gyms.results.BenchmarkResults.__init__`"
     },
     {
       "line": 158,
       "column": 35,
       "stop_line": 158,
       "stop_column": 54,
-      "path": "src/physicalai/benchmark/benchmark.py",
+      "path": "src/physicalai/benchmark/gyms/benchmark.py",
       "code": -2,
       "name": "missing-argument",
-      "description": "Missing argument `mean_reward` in function `physicalai.benchmark.results.BenchmarkResults.__init__`",
-      "concise_description": "Missing argument `mean_reward` in function `physicalai.benchmark.results.BenchmarkResults.__init__`",
-      "severity": "error"
+      "description": "Missing argument `mean_reward` in function `physicalai.benchmark.gyms.results.BenchmarkResults.__init__`",
+      "concise_description": "Missing argument `mean_reward` in function `physicalai.benchmark.gyms.results.BenchmarkResults.__init__"
     },
     {
       "line": 297,
       "column": 40,
       "stop_line": 297,
       "stop_column": 46,
-      "path": "src/physicalai/benchmark/benchmark.py",
+      "path": "src/physicalai/benchmark/gyms/benchmark.py",
       "code": -2,
       "name": "bad-argument-type",
       "description": "Argument `InferenceModel | Policy` is not assignable to parameter `policy` with type `Policy` in function `_get_policy_name`",
@@ -41,7 +39,7 @@
       "column": 13,
       "stop_line": 426,
       "stop_column": 26,
-      "path": "src/physicalai/benchmark/benchmark.py",
+      "path": "src/physicalai/benchmark/gyms/benchmark.py",
       "code": -2,
       "name": "bad-override-param-name",
       "description": "Class member `InferenceModelPolicyWrapper.select_action` overrides parent class `Policy` in an inconsistent manner\n  Got parameter name `observation`, expected `batch`",
@@ -53,7 +51,7 @@
       "column": 42,
       "stop_line": 70,
       "stop_column": 63,
-      "path": "src/physicalai/benchmark/results.py",
+      "path": "src/physicalai/benchmark/gyms/results.py",
       "code": -2,
       "name": "bad-typed-dict-key",
       "description": "`list[dict[str, Any]]` is not assignable to TypedDict key `per_episode_data` with type `float | int | str`",
@@ -65,7 +63,7 @@
       "column": 5,
       "stop_line": 118,
       "stop_column": 25,
-      "path": "src/physicalai/benchmark/results.py",
+      "path": "src/physicalai/benchmark/gyms/results.py",
       "code": -2,
       "name": "bad-class-definition",
       "description": "Dataclass field `overall_success_rate` without a default may not follow dataclass field with a default",
@@ -77,7 +75,7 @@
       "column": 5,
       "stop_line": 128,
       "stop_column": 16,
-      "path": "src/physicalai/benchmark/results.py",
+      "path": "src/physicalai/benchmark/gyms/results.py",
       "code": -2,
       "name": "bad-class-definition",
       "description": "Dataclass field `mean_reward` without a default may not follow dataclass field with a default",
@@ -89,7 +87,7 @@
       "column": 22,
       "stop_line": 292,
       "stop_column": 57,
-      "path": "src/physicalai/benchmark/results.py",
+      "path": "src/physicalai/benchmark/gyms/results.py",
       "code": -2,
       "name": "missing-argument",
       "description": "Missing argument `overall_success_rate` in function `BenchmarkResults.__init__`",
@@ -101,7 +99,7 @@
       "column": 22,
       "stop_line": 292,
       "stop_column": 57,
-      "path": "src/physicalai/benchmark/results.py",
+      "path": "src/physicalai/benchmark/gyms/results.py",
       "code": -2,
       "name": "missing-argument",
       "description": "Missing argument `mean_reward` in function `BenchmarkResults.__init__`",
