+++ library/src/physicalai/benchmark/gyms/libero/libero.py
@@ -22,7 +22,7 @@ from __future__ import annotations
 from enum import IntEnum
 from typing import TYPE_CHECKING
 
-from physicalai.benchmark.benchmark import Benchmark
+from physicalai.benchmark.gyms.benchmark import Benchmark
 
 if TYPE_CHECKING:
     from pathlib import Path
